Fix pretty stack tracing
